requirements.txt
typer[all]>=0.12
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0
litellm>=1.40
tiktoken>=0.7
pdfplumber>=0.11
python-docx>=1.1
markdown>=3.6
reportlab>=4.0
jinja2>=3.1
rich>=13.0
pytest>=8.0
